Change UserFragment to a Activity
