+++ app/src/main/res/layout/activity_challenge.xml
